ASSET Documentation

Theoretical Background

The method used in the ASSET package requires the following maps as inputs:

data maps $(d_\ell)_{\ell \in {1:L}}$, where $L$ is the amount of dithers/acquisitions/frames;

weights maps $w_\ell$, where each element can be computed as the inverse variance of the pixel, forming the matrix $W_{\ell}= \Sigma_{\ell}^{-2}$. We assume that a defective pixel or artifacts have an infinite variance, i.e. a zero entry in $W_{\ell}$;

spatial coordinate maps $X_\ell$ where $0$ should correspond to the center of the studied object;

spectral coordinate maps $\Lambda_\ell$.

The method outputs are the extracted spectrum $z$, sampled over a given regular wavelength grid $(\lambda_n)_{n \in 1:N}$, and the parameters $\theta$ of the fitted PSF model. They are obtained by solving:

$

z,\theta \in \mathrm{arg min} \Big{\sum\ell\Vert d\ell - (m\ell(z,\theta) + b)\Vert{W{\ell}}^2 + \muz\mathcal{R}z(z) + \mu\theta\mathcal{R}\theta(\theta) + \mub\mathcal{R}b(b), \Big} $ where \mathcal{R}z$, $\mathcal{R}_\theta$ and $\mathcal{R}_b$ are respectively the regularization of the extracted spectra, of the PSF parameters if required and of the background, with hyperparameters $\mu_z$, $\mu_\theta$ and $\mu_b$. The spatial distribution maps $X$ and the background map $b$ are auto-calibrated in the process. The model of the data $ m\ell(z,\theta) = \alpha\ell Z(\Lambda\ell,z) \odot H(\theta, X\ell, \Lambda\ell) $ is the Hadamard (element-wise) product of the spectrum interpolated in the camera plane $ Z(\Lambda,z){j,k}=\sumn\phi\Big(\frac{\Lambda{j,k}-\lambdan}{\delta\lambda}\Big)z_n $

with $\phi$ an interpolation kernel, and of the chromatic PSF $H$ (parametric or non-parametric). The package provide several ParametricPSF and NonParametricPSF and the users can easily implement their own. A ParametricPSF $H$ is a function parametrized by a few unknown variables $\theta_m$, thus with low degrees of freedom, e.g. a Gaussian chromatic with a minimum width model: $ H(\theta, X\ell, \Lambda\ell){j,k} = \big(2 \pi(\theta1 \Lambda{j,k\ell}^2 + \theta2)\big)^{-1}\exp\Big(-\frac{X{j,k,\ell}^2}{2(\theta1\Lambda{j,k,\ell}^2+\theta2)}\Big). $

It does not require any regularization. See @The:2023 and @Denneulin:2023 for more details.

A NonParametricPSF is parametrized directly by some profile $(\theta_m)_{m \in 1:M}$, thus with high degrees of freedom , e.g. taking $o$ order of the speckles expansion model [@Devaney:2017] which is the interpolation of the profiles $\theta_o$ in a reference plane of the spatial coordinates $(x_m)_{m \in 1:M}$,: $ H(\theta, X\ell, \Lambda\ell){j,k} = \sumo \gamma(\Lambda\ell)^o \summ \phi\Big(\frac{\gamma(\Lambda\ell)X\ell - xm}{\delta x}\Big)\theta{m,o}(x_m). $

For such a PSF, $\theta$ must be regularized. The hyperparameter is auto-calibrated in the method (see @The:2023 for and references therein for more details).

ASSET Documentation

Theoretical Background

ASSET.ChromaticSeriesExpansionsInterpolator — Type

struct ChromaticSeriesExpansionsInterpolator <: LinearMapping

A structure representing a chromatic series expansions interpolator, which performs linear mapping operations for chromatic data interpolation.

Arguments

cols: Tuple specifying the number of columns (output dimensions).

rows: Tuple specifying the number of rows (input dimensions).

ker: Kernel object used for interpolation.

X: Array of input positions.

Λ: Array of wavelength or chromatic values.

x: Range of positions for interpolation.

λref: Reference wavelength for normalization.

a: Parameter for scaling or regularization.

ChromaticSeriesExpansionsInterpolator{T}(ker::Kernel, X::AbstractArray{T,N}, Λ::AbstractArray{T,N}, x::AbstractRange; order=1, λref=maximum(Λ), a=0.)

Constructs a ChromaticSeriesExpansionsInterpolator with specified kernel, input arrays, and interpolation range. Optionally specify the interpolation order, reference wavelength, and scaling parameter.

Arguments

ker: Interpolation kernel.

X: Input position array.

Λ: Input wavelength/chromatic array.

x: Range of positions for interpolation.

order: Interpolation order (default: 1).

λref: Reference wavelength (default: maximum of Λ).

a: Scaling parameter (default: 0).

ChromaticSeriesExpansionsInterpolator(ker::Kernel, X::AbstractArray{<:Any,N}, Λ::AbstractArray{<:Any,N}, x::AbstractRange; order=1, λref=maximum(Λ), a=0.)

Constructs a ChromaticSeriesExpansionsInterpolator with type promotion to Float64 for the input arrays.

ASSET.CalibratedData — Type

struct CalibratedData{T<:AbstractFloat, N, D<:AbstractArray{T,N}}

A container for storing calibrated scientific data and associated weights w, the angular separation ρ_map and wavelength λ_map maps calibrating the detector.

Arguments

d::D: The main scientific data array.

w::D: The weight array corresponding to the data.

ρ_map::D: The spatial map associated with the data.

λ_map::D: The spectral map associated with the data.

The Base methods axes, size, eltype and show have been overload to be used with such a structure.

ASSET.CalibratedData

ASSET.ChromaticSeriesExpansionsInterpolator