ASSET Documentation

ASSET Documentation

Types

PSF models

Fitting methods

Theoretical Background

Types

ASSET.AbstractBkg — Type

AbstractBkg

Type that defines a background structure, that is used to specify a certain model of the background present in the data and its associated regularization. This structure is used when one want to both estimate the background sources of the data and extracting the spectrum of the object of interest.

To define a structure B of type AbstractBkg, the user needs to overload three methods:

julia> get_bkg(B)

which yields the background as an AbstractArray from parameters contained in B.

julia> regul(B)

which yields the result of the regularization term(s), applied to the curent background model, that are used in the a posteriori likelihood minimization.

julia> fit_bkg!(B, D)

where D is a CalibratedData. This last function is the one called by extract_spectrum to estimate the background when assuming the object spectrum known (alternate estimation).

See also

BkgMdl

ASSET.ParametricPSF — Type

ParametricPSF{N} <: AbstractPSF{N}

An abstract type representing a parametric point spread function (PSF) with N arguments. Subtypes P should implement specific parametric models for PSFs as well as parameters(::P) yielding its parameters, getfwhm(::P) its fwhm and ::P(::Int, ::Int) its value at a given position on the detector. Inherits from AbstractPSF{N}.

See Also

[AbstractPSF]

ASSET.NonParametricPSF — Type

NonParametricPSF{N} <: AbstractPSF{N}

An abstract type representing a non-parametric point spread function (PSF) with N arguments. Subtypes of NonParametricPSF are expected to define PSFs that are not described by a fixed set of parameters, but rather by data-driven or empirical representations. Subtypes P should implement specific parametric models for PSFs as well as parameters(::P) yielding its parameters, getfwhm(::P) its fwhm and ::P(::Int, ::Int) its value at a given position on the detector.This type is a subtype of AbstractPSF{N}.

See Also

[AbstractPSF]

ASSET.BkgMdl — Type

BkgMdl(b, R)

yields a structure of type AbstractBkg, composed of an AbstractArray b and an associated regularization R of type Regularization.

Example

# create a BkgMdl structure
julia> B = BkgMdl(b, R)
# add the structure to an AbstractArray
julia> A = ones(size(b))
julia> A + B

To apply the regularization to the background array:

julia> regul(B)

To fit the background to some CalibratedData D:

julia> fit_bkg!(B, D, true, kwds...)

where kwds specifies all the keywords used by the optimization method to estimate the background.

See also

AbstractBkg

ASSET.CalibratedData — Type

struct CalibratedData{T<:AbstractFloat, N, D<:AbstractArray{T,N}}

A container for storing calibrated scientific data and associated weights w, the angular separation ρ_map and wavelength λ_map maps calibrating the detector.

Arguments

d::D: The main scientific data array.

w::D: The weight array corresponding to the data.

ρ_map::D: The spatial map associated with the data.

λ_map::D: The spectral map associated with the data.

The Base methods axes, size, eltype and show have been overload to be used with such a structure.

PSF models

ASSET.chromGaussianPSF — Type

chromGaussianPSF(a) -> h

Yields an AbstractPSF h, more precisely a gaussian chromatic PSF parametrized by a, such that the gaussian variance is σ² = aλ².

Then:

h(ρ,λ)

gives the value of the gaussian chromatic PSF for:

ρ, a float giving the position in the slit (in pixels)

λ, a float giving the wavelength (in μm)

which is:

    exp(-1/2 ρ²/σ²)/sqrt(2πσ²)

ASSET.chromwmwGaussianPSF — Type

chromwmwGaussianPSF(p) -> h

Yields an AbstractPSF h, more precisely a gaussian chromatic PSF parametrized by two variables given in a vector p = [a,b]:

a the chromatic scalling

b the minimum width

such that the the variance of the PSF is σ² = aλ² + b

Then:

h(ρ,λ)

gives the value of the gaussian chromatic PSF for:

ρ, a float giving the position in the slit (in pixels)

λ, a float giving the wavelength (in μm)

which is:

    exp(-1/2 ρ²/σ²)/sqrt(2πσ²)

ASSET.chromMoffatPSF — Type

chromMoffatPSF(p) -> h

Yields an AbstractPSF h, more precisely a centrosymetric Moffat chromatic PSF parametrized by two variables given in a vector p=[a, β]:

`a the chromatic scalling

β the Moffat parameter

such that the the variance of the PSF is σ² = aλ²

Then:

h(ρ,λ)

gives the value of the centrosymetric Moffat chromatic PSF for:

ρ, a float giving the position in the slit (in pixels)

λ, a float giving the wavelength (in μm)

which is:

    (1 + ρ²/σ²)^(-β)

ASSET.chromwmwMoffatPSF — Type

chromwmwMoffatPSF(p) -> h

Yields an AbstractPSF h, more precisely a centrosymetric Moffat chromatic PSF parametrized by two variables given in a vector p=[a, b, β]:

a the chromatic scalling

b the minimum width

β the Moffat parameter

such that the the variance of the PSF is σ² = aλ² + b

Then:

h(ρ,λ)

gives the value of the centrosymetric Moffat chromatic PSF for:

ρ, a float giving the position in the slit (in pixels)

λ, a float giving the wavelength (in μm)

which is:

    (1 + ρ²/σ²)^(-β)

ASSET.SeriesExpansionPSF — Type

SeriesExpansionPSF(x) -> h

Yields an AbstractPSF h, more precisely a non-parametric PSF using a series expansion approach, parametrized by:

h: a matrix containing the series expansion coefficients

x: a range containing the wavelengths at which the series expansion coefficients are defined

a: a float giving the chromatic scaling (default is 0, i.e., no scaling)

ker: an interpolation kernel (default is CatmullRomSpline with Flat boundary conditions)

R: a regularization (default is Tikhonov regularization)

Then:

h(ρ,λ)

gives the value of the chromatic series expansion PSF for:

ρ, a float giving the position in the slit (in pixels)

λ, a float giving the wavelength (in μm)

which is given by the ChromaticSeriesExpansionsInterpolator structure.

See also

ChromaticSeriesExpansionsInterpolator

ASSET.ChromaticSeriesExpansionsInterpolator — Type

struct ChromaticSeriesExpansionsInterpolator <: LinearMapping

A structure representing a chromatic series expansions interpolator, which performs linear mapping operations for chromatic data interpolation.

Arguments

cols: Tuple specifying the number of columns (output dimensions).

rows: Tuple specifying the number of rows (input dimensions).

ker: Kernel object used for interpolation.

X: Array of input positions.

Λ: Array of wavelength or chromatic values.

x: Range of positions for interpolation.

λref: Reference wavelength for normalization.

a: Parameter for scaling or regularization.

ChromaticSeriesExpansionsInterpolator{T}(ker::Kernel, X::AbstractArray{T,N}, Λ::AbstractArray{T,N}, x::AbstractRange; order=1, λref=maximum(Λ), a=0.)

Constructs a ChromaticSeriesExpansionsInterpolator with specified kernel, input arrays, and interpolation range. Optionally specify the interpolation order, reference wavelength, and scaling parameter.

Arguments

ker: Interpolation kernel.

X: Input position array.

Λ: Input wavelength/chromatic array.

x: Range of positions for interpolation.

order: Interpolation order (default: 1).

λref: Reference wavelength (default: maximum of Λ).

a: Scaling parameter (default: 0).

ChromaticSeriesExpansionsInterpolator(ker::Kernel, X::AbstractArray{<:Any,N}, Λ::AbstractArray{<:Any,N}, x::AbstractRange; order=1, λref=maximum(Λ), a=0.)

Constructs a ChromaticSeriesExpansionsInterpolator with type promotion to Float64 for the input arrays.

Fitting methods

ASSET.extract_spectrum! — Function

extract_spectrum!(z, F, psf, D, Reg [, Bkg]; kwds...)

estimates the spectrum of an object observed with long-slit spectroscopy, when the data can be corrupted by a background component and noise. The direct model of such data can be writen: d = Bkg + Diag(H)*F*z + n where * the matrix product. z is the spectrum of the object of interest and F its associated interpolation operator, H the operator modeling the Point Spread Function (PSF) of the instrument, while Bkg is the background component to disentangle from the object of interest and n accounts for noises.

The data and there associated weights are given by the CalibratedData structure D, while the array H is produced using the method psf_map with the arguments psf and psf_center.

The estimation of the different unknowns relies on the minimization of the a posteriori likelihood, where Reg is the regularization function of z (of type InverseProblem.Regularization) and where Bkg is a structure of type AbstractBkg, which contains a way to produce the background component and its regularization function. If Bkg is not given, the algorithm will suppose that there is no background in the data and will only call fit_spectrum_and_psf!. If there is a Bkg given, the algorithm will alternate the estimation of the background using fit_bkg! with the extraction of the spectrum (done in fit_spectrum_and_psf!).

Finally, an auto-calibration step can be activated which will refine the parameters and center of the PSF.

Keywords

auto_calib : (Val(true) by default) precise if an auto-calibration step of the PSF must be done after extracting the spectrum. It can also take the value Val(:delay) which will run the algorithm without autocalib until it converges, before activating the `autocalib` and re-running the algorithm.

mask_width : (3 by default) defines the number of fwhm of the psf will be used to hide the object on the first iteration of the background estimtion (if thee is one).

max_iter : (1000 by default) defines the maximum number of iterations that can do the method (useful when auto_calib=true).

loss_tol : ((0,1e-6) by default) defines the absolute and relative tolerance between two consecutive iteration of the loss function as a stop criterion.

z_tol : ((0,1e-6) by default) defines the absolute and relative tolerance between two consecutive iteration of the estimate z as a stop criterion.

extract_kwds : ((verb=true,) by default) where other keywords can be given which are forwarded to the object_step method.

See also

AbstractBkg

[PointSpreadFunctions.AbstractPSF]

ASSET.fit_spectrum_and_psf! — Function

fit_spectrum_and_psf!(z, psf, F, D, Reg; kwds...)

yields the object spectrum z, the off-axis PSF psf, extracted from the CalibratedData D via an a posteriori likelihood minimization with regularization Reg. The model of the object is defined by Diag(H)*F*z where H is found via the psf arguments, using the method psf_map!. The optimization problem is solved by the vmlmb method defined in the OptimPackNextGen package by calling the method fit_spectrum!.

An auto-calibration step can be done to better estimate the parameters, inpsf, and shift to apply to the center of the PSF. The Bobyqa method of Powell is used to estimate these quantities.

Keywords

auto_calib : (Val(true) by default) precise if an auto-calibration step of the PSF must be done after extracting the spectrum.

psf_shift_bnds : (a vector of zero-values Tuple by default) defines the boundaries of the parameters of the PSF. If auto_calib=true, the user must specify them.

psf_center_bnds : (a zero-valued integer by default) defines the boundaries of the center of the PSF along the spectral axis. If auto_calib=true, the user must specify them.

max_iter : (1000 by default) defines the maximum number of iterations that can do the method (useful when auto_calib=true).

max_amors_iter : (10 by default) defines the maximum number of iterations that can do the method AMORS.

amors_tol : ((1e-3,1e-3) by default) defines the absolute and relative tolerance between two consecutive iteration of the 'AMORS' loss function as a stop criterion.

loss_tol : ((0,1e-6) by default) defines the absolute and relative tolerance between two consecutive iteration of the loss function as a stop criterion.

z_tol : ((0,1e-6) by default) defines the absolute and relative tolerance between two consecutive iteration of the estimate z as a stop criterion.

h_tol : ((0,1e-6) by default) defines the absolute and relative tolerance between two consecutive iteration of the estimate of the PSF as a stop criterion.

Other keywords can be given which are forwarded to the Bobyqa nethod.

Returns

z : the estimated object spectrum.

psf : the estimated off-axis PSF.

See also

[OptimPackNextGen.vmlbm]

[OptimPackNextGen.Powell.Bobyqa]

psf_map!

fit_spectrum!

[AMORS]

fit_spectrum_and_psf!(z, psf, F, D, Reg; kwds...)

yields the object spectrum z, the off-axis PSF psf, extracted from the CalibratedData D via an a posteriori likelihood minimization with regularization Reg. The model of the object is defined by Diag(H)*F*z where H is found via the psf arguments, using the method psf_map!. The optimization problem is solved by the vmlmb method defined in the OptimPackNextGen package by calling the method fit_spectrum!.

An auto-calibration step can be done to better estimate the parameters of the PSF. The Bobyqa method of Powell is used to estimate these quantities.

Keywords

auto_calib : (Val(true) by default) precise if an auto-calibration step of the PSF must be done after extracting the spectrum.

psf_params_bnds : (a vector of zero-values Tuple by default) defines the boundaries of the parameters of the PSF. If auto_calib=true, the user must specify them.

psf_center_bnds : (a zero-valued integer by default) defines the boundaries of the center of the PSF along the spectral axis. If auto_calib=true, the user must specify them.

max_iter : (1000 by default) defines the maximum number of iterations that can do the method (useful when auto_calib=true).

loss_tol : ((0,1e-6) by default) defines the absolute and relative tolerance between two consecutive iteration of the loss function as a stop criterion.

z_tol : ((0,1e-6) by default) defines the absolute and relative tolerance between two consecutive iteration of the estimate z as a stop criterion.

Other keywords can be given which are forwarded to the Bobyqa nethod.

See also

[OptimPackNextGen.vmlbm]

[OptimPackNextGen.Powell.Bobyqa]

psf_map!

fit_spectrum!

ASSET.fit_spectrum! — Function

fit_spectrum!(z, F, H, D, Reg ; kwds...)

check if the Regularization R can be used in a direct inversion framework before calling the solve function (either solve_analytic or solve_vmlmb). The keywords kwds are forwarded to the method that solves the problem.

See also

[InverseProblem.Regul]

ASSET.fit_psf_center! — Function

fit_psf_center(psf_center, psf, z, F, D; kwds...)

yields the center of the psf along the spectral axis, stored in psf_center, estimated by minimizing the likelihood formed with the CalibratedData D, the spectrum of the object z and the SparseInterpolator F.

#Keywords

psf_center_bnds : (a vector of zero-values Tuple by default) defines the boundaries of the center of the PSF along the spectral axis. The user must specify them.

rho_tol : (undef by default) defines the size of the trust-region used to estimate the parameters of the PSF.

Other keywords can be given which are forwarded to the Bobyqa nethod.

See also

[OptimPackNextGen.Powell.Bobyqa]

ASSET.fit_psf_params — Function

fit_psf_params(psf, psf_center, z, F, D; kwds...)

yields a new PSF structure of same type than psf, where its parameters are estimated by Powell's Bobyqa method as defined in OptimPackNextGen. The center of the PSF is defined by the vector psf_center, while the spectrum of the object of interest can be retireve using the vector z, the SparseInterpolatorF, and theCalibratedDataD`.

Keywords

psf_params_bnds : (a vector of zero-values Tuple by default) defines the boundaries of the parameters of the PSF. The user must specify them.

rho_tol : (undef by default) defines the size of the trust-region used to estimate the parameters of the PSF.

Other keywords can be given which are forwarded to the Bobyqa nethod.

See also

[OptimPackNextGen.Powell.Bobyqa]

ASSET.fit_psf_shift — Function

fitpsfshift(psf, z, F, D; psfshiftbnds = (0.,0.), kwds...)

Fits the shift parameter of a non-parametric point spread function (PSF) by minimizing a likelihood function.

Arguments

psf: The initial non-parametric PSF object.

z: The parameter vector for likelihood evaluation.

F: Sparse interpolator used in the likelihood calculation.

D: Calibrated data containing observed data, weights, and mapping information.

psf_shift_bnds: Optional. Tuple specifying the lower and upper bounds for the PSF shift parameter. Defaults to (0., 0.).

kwds...: Additional keyword arguments passed to the optimizer.

Returns

NonParametricPSF: A new PSF object with the optimized shift parameter.

ASSET.psf_map! — Function

psf_map!(map, h, ρ, λ)

store in the AbstractArray map the result of applying the psf function stored in the NonParametricPSF h, to each pixel i of the spatial and spectral maps ρ and λ.

See also

psf_map

psf_map!(map, h, ρ, λ)

store in the AbstractArray map the result of applying the psf function stored in the ParametricPSF h, to each pixel i of the spatial and spectral maps ρ and λ.

See also

psf_map

ASSET.psf_map — Function

psf_map(h, ρ, λ)

yield the result of psf_map! and store it in a new AbstractArray.

See also

psf_map!

psf_map(h, ρ, λ)

yield the result of psf_map! and store it in a new AbstractArray.

See also

psf_map!

ASSET.loss — Function

loss(D, H, F, z, Reg [, Bkg=undef])

yields the value of the loss function used to estimate the different parameters of the problem: (Diag(H)*F*z + Bkg.b - D.d)'*Diag(D.w)*(Diag(H)*F*z + Bkg.b - D.d) + Reg(z) + regul(Bkg) where Bkg.b and regul(Bkg) are not taken into account if Bkg=undef.

See also

AbstractBkg

[InverseProblem.Regul]

ASSET.solve_analytic — Function

solve_analytic!(F, H, D, Reg)

yields the estimator by directly inverting the Normal equations: (F'*H'*Diag(D.w)*H*F + get_grad_op(R)) * z = F'.H*Diag(D.w)*D.d where D is a CalibratedData, get_grad_op is a method returning the operator of the gradient of the Regularization Reg and z is the estimator the user is looking for.

See also

[InversePbm.get_grad_op]

ASSET.solve_vmlmb — Function

solve_vmlmb!(z0, A, D, Reg; [nonnegative=false,] kwds...)

uses the vmlmb! method of OptimPackNextGen to estimate the solution of: argmin (A*z - D.d)'*Diag(D.w)*(A*z - D.d) + Reg(z) z with D a CalibratedData. The result is stored in z0.

It is possible to indicate to the method vmlmb a positivity constraint for z by using the nonnegative keyword, as well as indicate more keywords to constrain the optimization.

See also

[OptimPackNextGen.vmlbm!]

ASSET.mask_object — Function

mask_object(ρ_map, λ_map, psf [; mask_width=3])

yields a mask array of same size than ρ_map and λ_map, where all the pixels of distance less than the fwhm of psf times the mask_width are flagged as zeros, while the rest are at unitary level.

To do so, the user needs to make sure that the ρ_map has its origin centered on the object of interest.

See also

[PointSpreadFunctions.AbstractPSF]

[PointSpreadFunctions.get_fwhm]

link to ASSET.AbstractBkg

link to ASSET.BkgMdl

link to ASSET.psf_map!

link to ASSET.psf_map

link to ASSET.fit_spectrum!

link to ASSET.ChromaticSeriesExpansionsInterpolator

ASSET.AbstractBkg

ASSET.BkgMdl

ASSET.CalibratedData

ASSET.ChromaticSeriesExpansionsInterpolator

ASSET.NonParametricPSF

ASSET.ParametricPSF

ASSET.SeriesExpansionPSF

ASSET.chromGaussianPSF

ASSET.chromMoffatPSF

ASSET.chromwmwGaussianPSF

ASSET.chromwmwMoffatPSF

ASSET.extract_spectrum!

ASSET.fit_psf_center!

ASSET.fit_psf_params

ASSET.fit_psf_shift

ASSET.fit_spectrum!

ASSET.fit_spectrum_and_psf!

ASSET.loss

ASSET.mask_object

ASSET.psf_map

ASSET.psf_map!

ASSET.solve_analytic

ASSET.solve_vmlmb

Theoretical Background

The method used in the ASSET package requires the following maps as inputs:

data maps $(d_\ell)_{\ell \in {1:L}}$, where $L$ is the amount of dithers/acquisitions/frames;

weights maps $w_\ell$, where each element can be computed as the inverse variance of the pixel, forming the matrix $W_{\ell}= \Sigma_{\ell}^{-2}$. We assume that a defective pixel or artifacts have an infinite variance, i.e. a zero entry in $W_{\ell}$;

spatial coordinate maps $X_\ell$ where $0$ should correspond to the center of the studied object;

spectral coordinate maps $\Lambda_\ell$.

The method outputs are the extracted spectrum $z$, sampled over a given regular wavelength grid $(\lambda_n)_{n \in 1:N}$, and the parameters $\theta$ of the fitted PSF model. They are obtained by solving:

\[ z,\theta \in \mathrm{arg min} \Big\{\sum_\ell\Vert d_\ell - (m_\ell(z,\theta) + b)\Vert_{W_{\ell}}^2 + \mu_z\mathcal{R}_z(z) + \mu_\theta\mathcal{R}_\theta(\theta) + \mu_b\mathcal{R}_b(b), \Big\} $ where $\mathcal{R}_z\]

, $\mathcal{R}_\theta$ and $\mathcal{R}_b$ are respectively the regularization of the extracted spectra, of the PSF parameters if required and of the background, with hyperparameters $\mu_z$, $\mu_\theta$ and $\mu_b$. The spatial distribution maps $X$ and the background map $b$ are auto-calibrated in the process. The model of the data

\[ m_\ell(z,\theta) = \alpha_\ell Z(\Lambda_\ell,z) \odot H(\theta, X_\ell, \Lambda_\ell)\]

is the Hadamard (element-wise) product of the spectrum interpolated in the camera plane

\[ Z(\Lambda,z)_{j,k}=\sum_n\phi\Big(\frac{\Lambda_{j,k}-\lambda_n}{\delta_\lambda}\Big)z_n $ with $\phi\]

an interpolation kernel, and of the chromatic PSF $H$ (parametric or non-parametric). The package provide several ParametricPSF and NonParametricPSF and the users can easily implement their own. A ParametricPSF $H$ is a function parametrized by a few unknown variables $\theta_m$, thus with low degrees of freedom, e.g. a Gaussian chromatic with a minimum width model:

\[ H(\theta, X_\ell, \Lambda_\ell)_{j,k} = \big(2 \pi(\theta_1 \Lambda_{j,k\ell}^2 + \theta_2)\big)^{-1}\exp\Big(-\frac{X_{j,k,\ell}^2}{2(\theta_1\Lambda_{j,k,\ell}^2+\theta_2)}\Big). $ It does not require any regularization. See @The:2023 and @Denneulin:2023 for more details. A `NonParametricPSF` is parametrized directly by some profile $(\theta_m)_{m \in 1:M}\]

, thus with high degrees of freedom , e.g. taking $o$ order of the speckles expansion model [@Devaney:2017] which is the interpolation of the profiles $\theta_o$ in a reference plane of the spatial coordinates $(x_m)_{m \in 1:M}$,:

\[ H(\theta, X_\ell, \Lambda_\ell)_{j,k} = \sum_o \gamma(\Lambda_\ell)^o \sum_m \phi\Big(\frac{\gamma(\Lambda_\ell)X_\ell - x_m}{\delta x}\Big)\theta_{m,o}(x_m). $ For such a PSF, $\theta\]

must be regularized. The hyperparameter is auto-calibrated in the method (see @The:2023 for and references therein for more details).